smol touchups
